+++ evtloop.c
@@ -50,6 +50,8 @@ static uint64_t radar39267328_fatal = 0;
 static uint64_t radar39623812 = 0;
 static uint64_t radar39623812_fatal = 0;
 static uint64_t radar42770257_fatal = 0;
+static uint64_t radar42783724 = 0;
+static uint64_t radar42783724_fatal = 0;
 static uint64_t missingtoken = 0;
 static uint64_t ooms = 0;
 
@@ -168,6 +170,7 @@ errout:
  * 39267328: audit(4): target path not resolved for rename(2)
  * 39623812: audit(4): path not resolved for utimes(2)
  * 42770257: audit(4): only one instead of four path tokens for renameat(2)
+ * 42783724: audit(4): target path not resolved for link(2)
  */
 #define TOKEN_ASSERT(EVENT, TOKEN, COND) \
 	if (!(COND)) { \
@@ -533,8 +536,6 @@ auef_readable(UNUSED int fd, void *udata) {
 	case AUE_RENAMEAT:
 	case AUE_LINK:
 	case AUE_LINKAT:
-	case AUE_SYMLINK:
-	case AUE_SYMLINKAT:
 		if (!LOGEVT_WANT(cfg->events, LOGEVT_FILEMON))
 			break;
 		TOKEN_ASSERT("rename", "return", ev.return_present);
@@ -544,10 +545,11 @@ auef_readable(UNUSED int fd, void *udata) {
 		}
 		TOKEN_ASSERT("rename", "subject", ev.subject_present);
 		/*
-		 * On at least 10.11.6, records include only an unresolved
-		 * target path.
+		 * On at least 10.11.6, AUE_RENAME and AUE_LINK records
+		 * include only an unresolved target path.
 		 *
-		 * Reported to Apple as radar 39267328 on 2018-04-08.
+		 * Reported to Apple as radar 39267328 on 2018-04-08 and
+		 * radar 42783724 on 2018-07-31 respectively.
 		 */
 		if (ev.path[1] && !ev.path[2]) {
 			/* two path tokens, assume both resolved */
@@ -563,6 +565,8 @@ auef_readable(UNUSED int fd, void *udata) {
 			/* three path tokens, assume third unresolved dpath */
 			if (ev.type == AUE_RENAME) {
 				radar39267328++;
+			} else if (ev.type == AUE_LINK) {
+				radar42783724++;
 			} else {
 				missingtoken++;
 				DEBUG(cfg->debug, "missingtoken",
@@ -577,10 +581,20 @@ auef_readable(UNUSED int fd, void *udata) {
 			if (!path) {
 				if (errno == ENOMEM)
 					ooms++;
-				else {
+				else if (ev.type == AUE_RENAME) {
 					radar39267328_fatal++;
 					DEBUG(cfg->debug,
 					      "radar39267328_fatal",
+					      "path[2]=%s "
+					      "pid=%i "
+					      "procmon_getcwd(pid)=>%s",
+					      ev.path[2],
+					      ev.subject.pid,
+					      cwd);
+				} else if (ev.type == AUE_LINK) {
+					radar42783724_fatal++;
+					DEBUG(cfg->debug,
+					      "radar42783724_fatal",
 					      "path[2]=%s "
 					      "pid=%i "
 					      "procmon_getcwd(pid)=>%s",
@@ -617,6 +631,10 @@ auef_readable(UNUSED int fd, void *udata) {
 			break;
 		filemon_touched(&ev.tv, &ev.subject, path);
 		break;
+
+	case AUE_SYMLINK:
+	case AUE_SYMLINKAT:
+		// XXX
 
 	/*
 	 * Unhandled events.
@@ -680,6 +698,8 @@ evtloop_stats(evtloop_stat_t *st) {
 	st->el_radar39267328 = radar39267328;
 	st->el_radar42770257 = radar42770257_fatal;
 	st->el_radar42770257_fatal = radar42770257_fatal;
+	st->el_radar42783724 = radar42783724;
+	st->el_radar42783724_fatal = radar42783724_fatal;
 	st->el_missingtoken = missingtoken;
 	st->el_ooms = ooms;
 	aupipe_stats(fileno(auef), &st->ap);
@@ -709,7 +729,8 @@ siginfo_arrived(UNUSED int sig, UNUSED void *udata) {
 	                "r38845784:0/%"PRIu64" "
 	                "r39267328:%"PRIu64"/%"PRIu64" "
 	                "r39623812:%"PRIu64"/%"PRIu64" "
-	                "r42770257:%"PRIu64"/%"PRIu64"\n",
+	                "r42770257:%"PRIu64"/%"PRIu64" "
+	                "r42783724:%"PRIu64"/%"PRIu64"\n",
 	                st.el_aupclobbers,
 	                st.el_aueunknowns,
 	                st.el_failedsyscalls,
@@ -723,7 +744,9 @@ siginfo_arrived(UNUSED int sig, UNUSED void *udata) {
 	                st.el_radar39623812_fatal,
 	                st.el_radar39623812,
 	                st.el_radar42770257_fatal,
-	                st.el_radar42770257);
+	                st.el_radar42770257,
+	                st.el_radar42783724_fatal,
+	                st.el_radar42783724);
 
 	fprintf(stderr, "procmon "
 	                "actprc:%"PRIu32" "
@@ -989,6 +1012,8 @@ evtloop_run(config_t *cfg) {
 	radar39623812_fatal = 0;
 	radar39623812 = 0;
 	radar42770257_fatal = 0;
+	radar42783724 = 0;
+	radar42783724_fatal = 0;
 	missingtoken = 0;
 	ooms = 0;
 	xnumon_pid = getpid();