+++ NEWS.md
@@ -5,12 +5,15 @@
 -   Detect a bug in audit(4) where calls to renameat(2) sometimes lead to
     `AUE_RENAMEAT` records with only one instead of four path tokens
     (radar 42770257).
+-   Detect a bug in audit(4) where calls to link(2) always lead to
+    `AUE_LINK` records with only three instead of four path tokens
+    (radar 42783724).
 
 Event schema changes:
 
 -   Event schema version increased to 7.  Changes affect eventcode 1.
--   Eventcode 1 added `evtloop.radar42770257` and
-    `evtloop.radar42770257_fatal`.
+-   Eventcode 1 added `evtloop.radar42770257`, `evtloop.radar42770257_fatal`,
+    `evtloop.radar42783724` and `evtloop.radar42783724_fatal`.
 
 ---